requirements.txt
streamlit
supabase-py
pandas
openpyxl
plotly
python-dotenv